refectoring api
