+++ api/templates/cc-by-sa.txt
