+++ api/templates/cc-by.txt
