+++ shared/__version__.py
